move streaming tests to integration
