--- tests/test_streaming_sync.py
@@ -1,696 +1,0 @@
-# tests/test_streaming_fanin_metrics.py
-"""
-Tests for streaming/async fan-in behavior and metric accounting.
-
-This suite validates:
-  1) Early start of downstream when `start_when=first_batch`.
-  2) Delayed start of downstream when no `start_when` hint is provided.
-  3) Multi-stream fan-in to a single downstream consumer.
-  4) Exactness and isolation of metric aggregation.
-  5) Idempotency of metrics on duplicated STATUS events (BATCH_OK / TASK_DONE).
-
-All tests rely on in-memory Kafka and DB, configured via fixtures from `conftest.py`.
-"""
-
-from __future__ import annotations
-
-import asyncio
-import inspect
-from typing import Any
-
-import pytest
-
-from flowkit.core.log import log_context
-from tests.helpers import BROKER, AIOKafkaProducerMock
-from tests.helpers.graph import (
-    make_graph,
-    node_by_id,
-    prime_graph,
-    wait_node_finished,
-    wait_node_not_running_for,
-    wait_node_running,
-    wait_task_finished,
-)
-from tests.helpers.handlers import build_analyzer_handler, build_indexer_handler
-
-# Limit worker types for this module (see conftest._worker_types_from_marker).
-pytestmark = pytest.mark.worker_types("indexer,enricher,ocr,analyzer")
-
-
-# ───────────────────────── Small helpers ─────────────────────────
-
-
-def _node_status(doc: dict[str, Any], node_id: str) -> Any:
-    """Get node status by node_id from a task doc."""
-    n = node_by_id(doc, node_id)
-    return n.get("status") if n else None
-
-
-def _make_indexer(node_id: str, total: int, batch: int) -> dict[str, Any]:
-    """
-    Build a simple indexer node:
-      - emits `total` items in batches of `batch`.
-      - uses inline input to avoid external dependencies.
-    """
-    return {
-        "node_id": node_id,
-        "type": "indexer",
-        "depends_on": [],
-        "fan_in": "all",
-        "io": {"input_inline": {"batch_size": batch, "total_skus": total}},
-        "status": None,
-        "attempt_epoch": 0,
-    }
-
-
-def _get_count(doc: dict[str, Any], node_id: str) -> int:
-    """Return the aggregated 'count' metric from a node's stats (0 if absent)."""
-    node = node_by_id(doc or {}, node_id)
-    return int((node.get("stats") or {}).get("count") or 0)
-
-
-# ───────────────────────── Tests ─────────────────────────
-
-
-@pytest.mark.asyncio
-async def test_start_when_first_batch_starts_early(env_and_imports, inmemory_db, coord, worker_factory, tlog):
-    """
-    Downstream should start while upstream is still running when `start_when=first_batch` is set.
-    """
-    cd, _ = env_and_imports
-    tlog.debug("test.start", event="test.start", test_name="start_when_first_batch_starts_early")
-
-    # Start workers via factory (separate handler instances, shared role topics).
-    await worker_factory(
-        ("indexer", build_indexer_handler(db=inmemory_db)),
-        ("analyzer", build_analyzer_handler(db=inmemory_db)),
-    )
-
-    u = _make_indexer("u", total=12, batch=4)  # 3 upstream batches
-    d = {
-        "node_id": "d",
-        "type": "analyzer",
-        "depends_on": ["u"],
-        "fan_in": "all",
-        "io": {
-            "start_when": "first_batch",
-            "input_inline": {
-                "input_adapter": "pull.from_artifacts",
-                "input_args": {"from_nodes": ["u"], "poll_ms": 30},
-            },
-        },
-        "status": None,
-        "attempt_epoch": 0,
-    }
-
-    graph = make_graph(nodes=[u, d], edges=[("u", "d")])
-    graph = prime_graph(cd, graph)
-
-    task_id = await coord.create_task(params={}, graph=graph)
-    tlog.debug("task.created", event="task.created", test_name="start_when_first_batch_starts_early", task_id=task_id)
-
-    with log_context(task_id=task_id):
-        doc_when_d_runs = await wait_node_running(inmemory_db, task_id, "d", timeout=6.0)
-        st_u = _node_status(doc_when_d_runs, "u")
-        tlog.debug(
-            "firstbatch.start_observed",
-            event="firstbatch.start_observed",
-            upstream=str(st_u),
-            downstream="running",
-        )
-
-        # Upstream must not yet be finished; otherwise early start is not verified.
-        assert not str(st_u).endswith("finished"), "Upstream already finished, early start not verified"
-
-        # Eventually everything must finish.
-        tdoc = await wait_task_finished(inmemory_db, task_id, timeout=12.0)
-        final = {n["node_id"]: str(n["status"]) for n in tdoc["graph"]["nodes"]}
-        tlog.debug("firstbatch.final", event="firstbatch.final", statuses=final)
-        assert node_by_id(tdoc, "u")["status"] == cd.RunState.finished
-        assert node_by_id(tdoc, "d")["status"] == cd.RunState.finished
-
-
-@pytest.mark.asyncio
-async def test_after_upstream_complete_delays_start(
-    env_and_imports, inmemory_db, coord, worker_factory, monkeypatch, tlog
-):
-    """
-    Without start_when, the downstream must not start until the upstream is fully finished.
-    We first assert a negative hold window while the upstream is running, then assert the
-    downstream starts and finishes after the upstream completes.
-    """
-    cd, _ = env_and_imports
-    tlog.debug("test.start", event="test.start", test_name="after_upstream_complete_delays_start")
-
-    # Slow down indexer batch processing: 2 batches x ~0.45s ~= 0.9s total
-    monkeypatch.setenv("TEST_IDX_PROCESS_SLEEP_SEC", "0.45")
-
-    # Start workers
-    await worker_factory(
-        ("indexer", build_indexer_handler(db=inmemory_db)),
-        ("analyzer", build_analyzer_handler(db=inmemory_db)),
-    )
-
-    # u: emits 10 SKUs in 2 batches of 5; d depends on u; no start_when
-    u = _make_indexer("u", total=10, batch=5)
-    d = {
-        "node_id": "d",
-        "type": "analyzer",
-        "depends_on": ["u"],
-        "fan_in": "all",  # wait for u to fully complete
-        "io": {
-            "input_inline": {
-                "input_adapter": "pull.from_artifacts",
-                "input_args": {"from_nodes": ["u"], "poll_ms": 30},
-            }
-        },
-        "status": None,
-        "attempt_epoch": 0,
-    }
-
-    graph = prime_graph(cd, make_graph(nodes=[u, d], edges=[("u", "d")]))
-    task_id = await coord.create_task(params={}, graph=graph)
-    tlog.debug("task.created", event="task.created", test_name="after_upstream_complete_delays_start", task_id=task_id)
-
-    with log_context(task_id=task_id):
-        # 1) While u is still running (~0.9s), d must NOT enter running (hold < u duration).
-        await wait_node_not_running_for(inmemory_db, task_id, "d", hold=0.5)
-
-        # 2) Wait until u finished.
-        await wait_node_finished(inmemory_db, task_id, "u", timeout=3.0)
-
-        # 3) Now d should start and quickly finish.
-        await wait_node_finished(inmemory_db, task_id, "d", timeout=2.0)
-        tlog.debug("delayed.start.ok", event="delayed.start.ok")
-
-
-@pytest.mark.asyncio
-async def test_multistream_fanin_stream_to_one_downstream(env_and_imports, inmemory_db, coord, worker_factory, tlog):
-    """
-    Multi-stream fan-in: three upstream indexers stream into one analyzer.
-    Analyzer should start early on first batch and eventually see the full flow.
-    """
-    cd, _ = env_and_imports
-    tlog.debug("test.start", event="test.start", test_name="multistream_fanin_stream_to_one_downstream")
-
-    await worker_factory(
-        ("indexer", build_indexer_handler(db=inmemory_db)),
-        ("indexer", build_indexer_handler(db=inmemory_db)),
-        ("indexer", build_indexer_handler(db=inmemory_db)),
-        ("analyzer", build_analyzer_handler(db=inmemory_db)),
-    )
-
-    # Three sources with several batches each.
-    u1 = _make_indexer("u1", total=9, batch=3)  # 3 batches
-    u2 = _make_indexer("u2", total=8, batch=4)  # 2 batches
-    u3 = _make_indexer("u3", total=12, batch=3)  # 4 batches
-
-    d = {
-        "node_id": "d",
-        "type": "analyzer",
-        "depends_on": ["u1", "u2", "u3"],
-        "fan_in": "any",
-        "io": {
-            "start_when": "first_batch",
-            "input_inline": {
-                "input_adapter": "pull.from_artifacts",
-                "input_args": {"from_nodes": ["u1", "u2", "u3"], "poll_ms": 25},
-            },
-        },
-        "status": None,
-        "attempt_epoch": 0,
-    }
-
-    graph = make_graph(nodes=[u1, u2, u3, d], edges=[("u1", "d"), ("u2", "d"), ("u3", "d")], agg={"after": "d"})
-    graph = prime_graph(cd, graph)
-
-    task_id = await coord.create_task(params={}, graph=graph)
-    tlog.debug(
-        "task.created", event="task.created", test_name="multistream_fanin_stream_to_one_downstream", task_id=task_id
-    )
-
-    with log_context(task_id=task_id):
-        doc_when_d_runs = await wait_node_running(inmemory_db, task_id, "d", timeout=8.0)
-        st_u = {nid: str(_node_status(doc_when_d_runs, nid)) for nid in ("u1", "u2", "u3")}
-        tlog.debug(
-            "multistream.start_observed", event="multistream.start_observed", upstream=st_u, downstream="running"
-        )
-
-        # Not all upstreams should be finished at the analyzer start moment.
-        assert sum(1 for s in st_u.values() if s.endswith("finished")) < 3
-
-        # Final state must be finished for all nodes.
-        tdoc = await wait_task_finished(inmemory_db, task_id, timeout=14.0)
-        final = {n["node_id"]: str(n["status"]) for n in tdoc["graph"]["nodes"]}
-        tlog.debug("multistream.final", event="multistream.final", statuses=final)
-
-        assert node_by_id(tdoc, "d")["status"] == cd.RunState.finished
-        assert node_by_id(tdoc, "u1")["status"] == cd.RunState.finished
-        assert node_by_id(tdoc, "u2")["status"] == cd.RunState.finished
-        assert node_by_id(tdoc, "u3")["status"] == cd.RunState.finished
-
-        # Analyzer aggregated count should be at least the total items across sources.
-        got = _get_count(tdoc, "d")
-        tlog.debug("multistream.count", event="multistream.count", count=got)
-        assert got >= (9 + 8 + 12)
-
-
-@pytest.mark.asyncio
-async def test_metrics_single_stream_exact_count(env_and_imports, inmemory_db, coord, worker_factory, tlog):
-    """
-    Single upstream -> single downstream: analyzer's aggregated 'count' must equal the total items.
-    """
-    cd, _ = env_and_imports
-    tlog.debug("test.start", event="test.start", test_name="metrics_single_stream_exact_count")
-
-    await worker_factory(
-        ("indexer", build_indexer_handler(db=inmemory_db)),
-        ("analyzer", build_analyzer_handler(db=inmemory_db)),
-    )
-
-    total, batch = 13, 5  # 5 + 5 + 3
-    u = _make_indexer("u", total=total, batch=batch)
-    d = {
-        "node_id": "d",
-        "type": "analyzer",
-        "depends_on": ["u"],
-        "fan_in": "all",
-        "io": {
-            "start_when": "first_batch",
-            "input_inline": {
-                "input_adapter": "pull.from_artifacts",
-                "input_args": {"from_nodes": ["u"], "poll_ms": 25},
-            },
-        },
-        "status": None,
-        "attempt_epoch": 0,
-    }
-    graph = make_graph(nodes=[u, d], edges=[("u", "d")], agg={"after": "d"})
-    graph = prime_graph(cd, graph)
-
-    task_id = await coord.create_task(params={}, graph=graph)
-    tdoc = await wait_task_finished(inmemory_db, task_id, timeout=12.0)
-
-    got = _get_count(tdoc, "d")
-    tlog.debug("metrics.single", event="metrics.single", expect=total, got=got)
-    assert got == total
-
-
-@pytest.mark.asyncio
-async def test_metrics_multistream_exact_sum(env_and_imports, inmemory_db, coord, worker_factory, tlog):
-    """
-    Three upstreams -> one downstream: analyzer's aggregated 'count' must equal the sum of all totals.
-    """
-    cd, _ = env_and_imports
-    tlog.debug("test.start", event="test.start", test_name="metrics_multistream_exact_sum")
-
-    await worker_factory(
-        ("indexer", build_indexer_handler(db=inmemory_db)),
-        ("indexer", build_indexer_handler(db=inmemory_db)),
-        ("indexer", build_indexer_handler(db=inmemory_db)),
-        ("analyzer", build_analyzer_handler(db=inmemory_db)),
-    )
-
-    totals = {"u1": 6, "u2": 7, "u3": 9}  # sum = 22
-    u1 = _make_indexer("u1", total=totals["u1"], batch=3)
-    u2 = _make_indexer("u2", total=totals["u2"], batch=4)
-    u3 = _make_indexer("u3", total=totals["u3"], batch=4)
-
-    d = {
-        "node_id": "d",
-        "type": "analyzer",
-        "depends_on": ["u1", "u2", "u3"],
-        "fan_in": "any",
-        "io": {
-            "start_when": "first_batch",
-            "input_inline": {
-                "input_adapter": "pull.from_artifacts",
-                "input_args": {"from_nodes": ["u1", "u2", "u3"], "poll_ms": 20},
-            },
-        },
-        "status": None,
-        "attempt_epoch": 0,
-    }
-    graph = make_graph(nodes=[u1, u2, u3, d], edges=[("u1", "d"), ("u2", "d"), ("u3", "d")], agg={"after": "d"})
-    graph = prime_graph(cd, graph)
-
-    task_id = await coord.create_task(params={}, graph=graph)
-    tdoc = await wait_task_finished(inmemory_db, task_id, timeout=14.0)
-
-    expect = sum(totals.values())
-    got = _get_count(tdoc, "d")
-    tlog.debug("metrics.multi", event="metrics.multi", expect=expect, got=got)
-    assert got == expect
-
-
-@pytest.mark.asyncio
-async def test_metrics_partial_batches_exact_count(env_and_imports, inmemory_db, coord, worker_factory, tlog):
-    """
-    With a remainder in the last upstream batch, analyzer's 'count' must still exactly equal total.
-    """
-    cd, _ = env_and_imports
-    tlog.debug("test.start", event="test.start", test_name="metrics_partial_batches_exact_count")
-
-    await worker_factory(
-        ("indexer", build_indexer_handler(db=inmemory_db)),
-        ("analyzer", build_analyzer_handler(db=inmemory_db)),
-    )
-
-    total, batch = 10, 3  # 3 + 3 + 3 + 1
-    u = _make_indexer("u", total=total, batch=batch)
-    d = {
-        "node_id": "d",
-        "type": "analyzer",
-        "depends_on": ["u"],
-        "fan_in": "all",
-        "io": {
-            "start_when": "first_batch",
-            "input_inline": {
-                "input_adapter": "pull.from_artifacts",
-                "input_args": {"from_nodes": ["u"], "poll_ms": 25},
-            },
-        },
-        "status": None,
-        "attempt_epoch": 0,
-    }
-    graph = make_graph(nodes=[u, d], edges=[("u", "d")], agg={"after": "d"})
-    graph = prime_graph(cd, graph)
-
-    task_id = await coord.create_task(params={}, graph=graph)
-    tdoc = await wait_task_finished(inmemory_db, task_id, timeout=12.0)
-
-    got = _get_count(tdoc, "d")
-    tlog.debug("metrics.partial", event="metrics.partial", expect=total, got=got)
-    assert got == total
-
-
-@pytest.mark.asyncio
-async def test_metrics_isolation_between_tasks(env_and_imports, inmemory_db, coord, worker_factory, tlog):
-    """
-    Two back-to-back tasks must keep metric aggregation isolated per task document.
-    """
-    cd, _ = env_and_imports
-    tlog.debug("test.start", event="test.start", test_name="metrics_isolation_between_tasks")
-
-    await worker_factory(
-        ("indexer", build_indexer_handler(db=inmemory_db)),
-        ("analyzer", build_analyzer_handler(db=inmemory_db)),
-    )
-
-    async def run_once(total: int, batch: int):
-        u = _make_indexer("u", total=total, batch=batch)
-        d = {
-            "node_id": "d",
-            "type": "analyzer",
-            "depends_on": ["u"],
-            "fan_in": "all",
-            "io": {
-                "start_when": "first_batch",
-                "input_inline": {
-                    "input_adapter": "pull.from_artifacts",
-                    "input_args": {"from_nodes": ["u"], "poll_ms": 25},
-                },
-            },
-            "status": None,
-            "attempt_epoch": 0,
-        }
-        g = make_graph(nodes=[u, d], edges=[("u", "d")], agg={"after": "d"})
-        g = prime_graph(cd, g)
-        tid = await coord.create_task(params={}, graph=g)
-        with log_context(task_id=tid):
-            tdoc = await wait_task_finished(inmemory_db, tid, timeout=12.0)
-            return tid, _get_count(tdoc, "d")
-
-    tid1, c1 = await run_once(7, 3)
-    tid2, c2 = await run_once(11, 4)
-
-    tlog.debug("metrics.isolation", event="metrics.isolation", task1=tid1, count1=c1, task2=tid2, count2=c2)
-    assert c1 == 7
-    assert c2 == 11
-
-
-@pytest.mark.asyncio
-async def test_metrics_idempotent_on_duplicate_status_events(
-    env_and_imports, inmemory_db, coord, worker_factory, monkeypatch, tlog
-):
-    """
-    Duplicate STATUS events (BATCH_OK / TASK_DONE) must not double-increment aggregated metrics.
-
-    We monkeypatch AIOKafkaProducerMock.send_and_wait to produce the same STATUS event twice
-    for status topics. The Coordinator should deduplicate by envelope key and keep metrics stable.
-    """
-    cd, _ = env_and_imports
-    tlog.debug("test.start", event="test.start", test_name="metrics_idempotent_on_duplicate_status_events")
-
-    await worker_factory(
-        ("indexer", build_indexer_handler(db=inmemory_db)),
-        ("analyzer", build_analyzer_handler(db=inmemory_db)),
-    )
-
-    # Save the original method to call once, then produce duplicate via broker.
-    orig_send = AIOKafkaProducerMock.send_and_wait
-
-    async def dup_status(self, topic, value, key=None):
-        # First, the normal send.
-        await orig_send(self, topic, value, key)
-        # Then, if it's a STATUS event, produce a duplicate of the same envelope.
-        if topic.startswith("status.") and (value or {}).get("msg_type") == "event":
-            kind = (value.get("payload") or {}).get("kind") or ""
-            if kind in ("BATCH_OK", "TASK_DONE"):
-                await BROKER.produce(topic, value)
-
-    # Patch the mock implementation used by both worker and coordinator sides.
-    monkeypatch.setattr("tests.helpers.kafka.AIOKafkaProducerMock.send_and_wait", dup_status, raising=True)
-
-    total, batch = 18, 6  # 3 upstream batches
-    u = _make_indexer("u", total=total, batch=batch)
-    d = {
-        "node_id": "d",
-        "type": "analyzer",
-        "depends_on": ["u"],
-        "fan_in": "all",
-        "io": {
-            "start_when": "first_batch",
-            "input_inline": {
-                "input_adapter": "pull.from_artifacts",
-                "input_args": {"from_nodes": ["u"], "poll_ms": 25},
-            },
-        },
-        "status": None,
-        "attempt_epoch": 0,
-    }
-    graph = make_graph(nodes=[u, d], edges=[("u", "d")], agg={"after": "d"})
-    graph = prime_graph(cd, graph)
-
-    task_id = await coord.create_task(params={}, graph=graph)
-    with log_context(task_id=task_id):
-        tdoc = await wait_task_finished(inmemory_db, task_id, timeout=12.0)
-        got = _get_count(tdoc, "d")
-        tlog.debug("metrics.dedup", event="metrics.dedup", expect=total, got=got)
-        assert got == total
-
-
-@pytest.mark.asyncio
-async def test_status_out_of_order_does_not_break_aggregation(
-    env_and_imports, inmemory_db, coord, worker_factory, monkeypatch, tlog
-):
-    """
-    BATCH_OK STATUS events arrive out of order → the aggregated metric remains correct.
-    We hold the first BATCH_OK so later ones arrive first.
-    """
-    cd, _ = env_and_imports
-    tlog.debug("test.start", event="test.start", test_name="status_out_of_order_does_not_break_aggregation")
-
-    # Slightly slow batches so TASK_DONE arrives after the delayed BATCH_OK.
-    monkeypatch.setenv("TEST_IDX_PROCESS_SLEEP_SEC", "0.2")
-
-    await worker_factory(
-        ("indexer", build_indexer_handler(db=inmemory_db)),
-        ("analyzer", build_analyzer_handler(db=inmemory_db)),
-    )
-
-    # Intercept producer: hold the first BATCH_OK and re-inject it via the broker later.
-    orig_send = AIOKafkaProducerMock.send_and_wait
-    state = {"held_once": False}
-
-    async def out_of_order(self, topic, value, key=None):
-        payload = (value or {}).get("payload") or {}
-        if (
-            topic.startswith("status.")
-            and (value or {}).get("msg_type") == "event"
-            and payload.get("kind") == "BATCH_OK"
-            and not state["held_once"]
-        ):
-            state["held_once"] = True
-
-            async def later():
-                await asyncio.sleep(0.15)
-                await BROKER.produce(topic, value)
-
-            state["delayed_task"] = asyncio.create_task(later())
-            state["delayed_task"].add_done_callback(lambda _t: state.pop("delayed_task", None))
-            return  # suppress original send; delayed one will be re-injected via BROKER
-
-        return await orig_send(self, topic, value, key)
-
-    monkeypatch.setattr("tests.helpers.kafka.AIOKafkaProducerMock.send_and_wait", out_of_order, raising=True)
-
-    total, batch = 12, 4  # 3 batches
-    u = _make_indexer("u", total=total, batch=batch)
-    d = {
-        "node_id": "d",
-        "type": "analyzer",
-        "depends_on": ["u"],
-        "fan_in": "all",
-        "io": {
-            "start_when": "first_batch",
-            "input_inline": {
-                "input_adapter": "pull.from_artifacts",
-                "input_args": {"from_nodes": ["u"], "poll_ms": 25},
-            },
-        },
-        "status": None,
-        "attempt_epoch": 0,
-    }
-    graph = make_graph(nodes=[u, d], edges=[("u", "d")], agg={"after": "d"})
-    graph = prime_graph(cd, graph)
-
-    task_id = await coord.create_task(params={}, graph=graph)
-    with log_context(task_id=task_id):
-        tdoc = await wait_task_finished(inmemory_db, task_id, timeout=12.0)
-        got = _get_count(tdoc, "d")
-        tlog.debug("metrics.ooo", event="metrics.ooo", expect=total, got=got)
-        assert got == total
-
-
-@pytest.mark.asyncio
-async def test_metrics_dedup_persists_across_coord_restart(
-    env_and_imports, inmemory_db, coord, worker_factory, monkeypatch, tlog
-):
-    """
-    Metric deduplication survives a coordinator restart: duplicates of STATUS
-    (BATCH_OK/TASK_DONE) sent during the restart must not double-count.
-    Skips if the fixture `coord` has no `restart` method`.
-    """
-    cd, _ = env_and_imports
-    tlog.debug("test.start", event="test.start", test_name="metrics_dedup_persists_across_coord_restart")
-
-    restart_fn = getattr(coord, "restart", None)
-    if restart_fn is None:
-        pytest.skip("Coordinator restart is not supported by this fixture")
-
-    await worker_factory(
-        ("indexer", build_indexer_handler(db=inmemory_db)),
-        ("analyzer", build_analyzer_handler(db=inmemory_db)),
-    )
-
-    orig_send = AIOKafkaProducerMock.send_and_wait
-    fired = {"restarted": False}
-
-    async def dup_and_restart(self, topic, value, key=None):
-        payload = (value or {}).get("payload") or {}
-        # (1) On the first BATCH_OK, restart the coordinator.
-        # (2) Always send the original.
-        # (3) Inject a duplicate STATUS to verify dedup after restart.
-        if (
-            not fired["restarted"]
-            and topic.startswith("status.")
-            and (value or {}).get("msg_type") == "event"
-            and payload.get("kind") == "BATCH_OK"
-        ):
-            fired["restarted"] = True
-            if inspect.iscoroutinefunction(restart_fn):
-                await restart_fn()  # soft restart
-            else:
-                restart_fn()
-
-        await orig_send(self, topic, value, key)
-        if topic.startswith("status.") and (value or {}).get("msg_type") == "event":
-            kind = (payload or {}).get("kind") or ""
-            if kind in ("BATCH_OK", "TASK_DONE"):
-                await BROKER.produce(topic, value)
-
-    monkeypatch.setattr("tests.helpers.kafka.AIOKafkaProducerMock.send_and_wait", dup_and_restart, raising=True)
-
-    total, batch = 18, 6  # 3 batches
-    u = _make_indexer("u", total=total, batch=batch)
-    d = {
-        "node_id": "d",
-        "type": "analyzer",
-        "depends_on": ["u"],
-        "fan_in": "all",
-        "io": {
-            "start_when": "first_batch",
-            "input_inline": {
-                "input_adapter": "pull.from_artifacts",
-                "input_args": {"from_nodes": ["u"], "poll_ms": 25},
-            },
-        },
-        "status": None,
-        "attempt_epoch": 0,
-    }
-    graph = make_graph(nodes=[u, d], edges=[("u", "d")], agg={"after": "d"})
-    graph = prime_graph(cd, graph)
-
-    task_id = await coord.create_task(params={}, graph=graph)
-    with log_context(task_id=task_id):
-        tdoc = await wait_task_finished(inmemory_db, task_id, timeout=14.0)
-        got = _get_count(tdoc, "d")
-        tlog.debug("metrics.dedup.restart", event="metrics.dedup.restart", expect=total, got=got)
-        assert got == total
-
-
-@pytest.mark.asyncio
-async def test_metrics_cross_talk_guard(env_and_imports, inmemory_db, coord, worker_factory, monkeypatch, tlog):
-    """
-    Two concurrent tasks (same node_ids across different task_ids) do not interfere:
-    final 'count' values are isolated per task.
-    """
-    cd, _ = env_and_imports
-    tlog.debug("test.start", event="test.start", test_name="metrics_cross_talk_guard")
-
-    # Slow down slightly to force overlap in execution.
-    monkeypatch.setenv("TEST_IDX_PROCESS_SLEEP_SEC", "0.1")
-
-    await worker_factory(
-        ("indexer", build_indexer_handler(db=inmemory_db)),
-        ("analyzer", build_analyzer_handler(db=inmemory_db)),
-    )
-
-    def build_graph(total: int, batch: int):
-        u = _make_indexer("u", total=total, batch=batch)
-        d = {
-            "node_id": "d",
-            "type": "analyzer",
-            "depends_on": ["u"],
-            "fan_in": "all",
-            "io": {
-                "start_when": "first_batch",
-                "input_inline": {
-                    "input_adapter": "pull.from_artifacts",
-                    "input_args": {"from_nodes": ["u"], "poll_ms": 20},
-                },
-            },
-            "status": None,
-            "attempt_epoch": 0,
-        }
-        g = make_graph(nodes=[u, d], edges=[("u", "d")], agg={"after": "d"})
-        return g
-
-    g1 = prime_graph(cd, build_graph(15, 5))
-    g2 = prime_graph(cd, build_graph(21, 4))
-
-    # Start both tasks back-to-back to ensure real overlap on topics.
-    t1 = await coord.create_task(params={}, graph=g1)
-    t2 = await coord.create_task(params={}, graph=g2)
-
-    with log_context(task_id=t1):
-        doc1 = await wait_task_finished(inmemory_db, t1, timeout=14.0)
-        c1 = _get_count(doc1, "d")
-    with log_context(task_id=t2):
-        doc2 = await wait_task_finished(inmemory_db, t2, timeout=14.0)
-        c2 = _get_count(doc2, "d")
-
-    tlog.debug("metrics.crosstalk", event="metrics.crosstalk", t1=t1, c1=c1, t2=t2, c2=c2)
-    assert c1 == 15
-    assert c2 == 21
